add extension load in extension module
