+++ demo-shell/src/assets/plugins/plugin1.json
@@ -10,7 +10,6 @@
   "$dependencies": [
     "@some/library1"
   ],
-
   "features": {
     "custom": {
       "key": "value"
