+++ lib/extensions/src/lib/services/app-extension.service.ts
@@ -16,12 +16,10 @@
  */
 
 import { Injectable } from '@angular/core';
-import {
-    ExtensionService,
-    ExtensionConfig,
-    ExtensionRef
-} from '@alfresco/adf-extensions';
+import { ExtensionConfig, ExtensionRef } from '../config/extension.config';
+import { ExtensionService } from '../services/extension.service';
 import { Observable, BehaviorSubject } from 'rxjs';
+import { LogService } from '@alfresco/adf-core';
 
 @Injectable({
     providedIn: 'root'
@@ -31,22 +29,22 @@ export class AppExtensionService {
 
     references$: Observable<ExtensionRef[]>;
 
-    constructor(private extensions: ExtensionService) {
+    constructor(private extensionService: ExtensionService, private logService: LogService) {
         this.references$ = this._references.asObservable();
     }
 
     async load() {
-        const config = await this.extensions.load();
+        const config = await this.extensionService.load();
         this.setup(config);
     }
 
     setup(config: ExtensionConfig) {
         if (!config) {
-            console.error('Extension configuration not found');
+            this.logService.error('Extension configuration not found');
             return;
         }
-        // tslint:disable-next-line:no-console
-        console.log('loaded extension config', config);
+
+        this.logService.log('loaded extension config', config);
 
         const references = (config.$references || [])
             .filter((entry) => typeof entry === 'object')
